+++ libs/server/AOF/Recover/RecoverLogDriver.cs
@@ -25,7 +25,7 @@ namespace Garnet.server
     /// <param name="untilAddress">End address in the append-only file for recovery.</param>
     /// <param name="untilSequenceNumber">Replay all records with sequence number to ensure prefix consistent recovery.</param>
     /// <param name="logger">Optional logger for diagnostic output.</param>
-    internal sealed class RecoverLogDriver(
+    internal sealed partial class RecoverLogDriver(
         AofProcessor aofProcessor,
         GarnetAppendOnlyFile appendOnlyFile,
         GarnetServerOptions serverOptions,
@@ -57,6 +57,16 @@ namespace Garnet.server
         /// permit can be stolen across tasks or leak across cycles.
         /// </summary>
         WorkReadyComplete[] signals = null;
+
+        /// <summary>
+        /// Set by a replay task when it encounters an entry whose sequence number exceeds
+        /// <see cref="untilSequenceNumber"/> (the prefix-consistency boundary). Because sequence numbers are
+        /// monotonically increasing, no further entries need to be applied once this is set. The leader
+        /// observes it only after every task has finished the current page (via <see cref="WorkReadyComplete.WaitCompleted"/>,
+        /// which establishes a happens-before), then cancels to stop handing subsequent pages — avoiding
+        /// wasted scans and, importantly, avoiding committing log metadata beyond the intended prefix.
+        /// </summary>
+        volatile bool prefixConsistencyBoundaryReached = false;
 
         /// <summary>
         /// Gets the total number of records that have been replayed.
@@ -130,8 +140,20 @@ namespace Garnet.server
                 // Wait for previous batch to complete before overwriting shared batch context
                 if (replayTasks != null)
                 {
+                    // Pass the cancellation token so a worker that cancelled/faulted (and therefore did NOT
+                    // signal completion, by design) unwinds this leader wait instead of deadlocking recovery.
                     for (var i = 0; i < signals.Length; i++)
-                        signals[i].WaitCompleted();
+                        _ = signals[i].WaitCompleted(cancellationToken: cts.Token);
+
+                    // Every task has now finished the previous page. If any of them reached the
+                    // prefix-consistency boundary, stop here so we neither scan further pages nor commit
+                    // log metadata beyond the intended prefix.
+                    if (prefixConsistencyBoundaryReached)
+                    {
+                        logger?.LogTrace("Reached prefix-consistency boundary (until sequence number {untilSequenceNumber}), stopping parallel replay", untilSequenceNumber);
+                        cts.Cancel();
+                        return;
+                    }
                 }
 
                 CreateAndRunIntraPageParallelReplayTasks();
@@ -150,7 +172,7 @@ namespace Garnet.server
                 if (nextAddress == untilAddress)
                 {
                     for (var i = 0; i < signals.Length; i++)
-                        signals[i].WaitCompleted();
+                        _ = signals[i].WaitCompleted(cancellationToken: cts.Token);
                     cts.Cancel();
                 }
             }
@@ -162,121 +184,7 @@ namespace Garnet.server
                 return;
 
             signals = [.. Enumerable.Range(0, serverOptions.AofReplayTaskCount).Select(_ => new WorkReadyComplete())];
-            replayTasks = [.. Enumerable.Range(0, serverOptions.AofReplayTaskCount).Select(i => Task.Run(() => ContinuousBackgroundReplayAsync(i, physicalSublog)))];
-        }
-
-        internal async Task ContinuousBackgroundReplayAsync(int replayTaskIdx, TsavoriteLog replaySublog)
-        {
-            var virtualSublogIdx = appendOnlyFile.GetVirtualSublogIdx(physicalSublogIdx, replayTaskIdx);
-            var signal = signals[replayTaskIdx];
-            while (!cts.Token.IsCancellationRequested)
-            {
-                try
-                {
-                    await signal.WaitReadyWorkAsync(cancellationToken: cts.Token).ConfigureAwait(false);
-                }
-                catch (TaskCanceledException) when (cts.IsCancellationRequested)
-                { }
-                catch (Exception ex)
-                {
-                    logger?.LogError(ex, "{method} failed at WaitAsync", nameof(ContinuousBackgroundReplayAsync));
-                    await cts.CancelAsync().ConfigureAwait(false);
-                    break;
-                }
-
-                if (cts.Token.IsCancellationRequested)
-                    break;
-
-                try
-                {
-                    unsafe
-                    {
-                        ReplayPage(
-                            replayTaskIdx,
-                            virtualSublogIdx,
-                            replaySublog,
-                            replayBatchContext.Record,
-                            replayBatchContext.RecordLength,
-                            replayBatchContext.CurrentAddress,
-                            replayBatchContext.IsProtected);
-                    }
-                }
-                catch (OperationCanceledException) when (cts.Token.IsCancellationRequested)
-                {
-                    // Cancelled during store disposal / prefix-consistency stop: break without signalling.
-                    break;
-                }
-                catch (Exception ex)
-                {
-                    logger?.LogError(ex, "{method} failed at replaying", nameof(ContinuousBackgroundReplayAsync));
-                    await cts.CancelAsync().ConfigureAwait(false);
-                    break;
-                }
-
-                // Signal completion ONLY after the page was fully applied. On cancellation or fault we
-                // break above without signalling, so the leader observes cancellation instead of a
-                // completion for a page that was not fully applied.
-                signal.SignalCompleted();
-            }
-        }
-
-        /// <summary>
-        /// Scans a single consumed page for the given replay task, applying every entry that this task
-        /// owns (per <see cref="AofProcessor.CanReplay"/>) exactly once. Invoked by each parallel replay
-        /// task (<see cref="ContinuousBackgroundReplayAsync"/>).
-        /// </summary>
-        private unsafe void ReplayPage(
-            int replayTaskIdx,
-            int virtualSublogIdx,
-            TsavoriteLog replaySublog,
-            byte* record,
-            int recordLength,
-            long currentAddress,
-            bool isProtected)
-        {
-            var ptr = record;
-            var maxSequenceNumber = 0L;
-
-            while (ptr < record + recordLength)
-            {
-                cts.Token.ThrowIfCancellationRequested();
-                var entryLength = appendOnlyFile.HeaderSize;
-                var payloadLength = replaySublog.UnsafeGetLength(ptr);
-                if (payloadLength > 0)
-                {
-                    var entryPtr = ptr + entryLength;
-                    var logAddressSequenceNumber = currentAddress + (ptr - record);
-                    Debug.Assert(logAddressSequenceNumber > 0, "Entry log address must be positive");
-                    // Check if entry is assigned for processing to this replay task and
-                    // the sequence number is below the threshold to ensure prefix consistency
-                    if (aofProcessor.CanReplay(entryPtr, replayTaskIdx, logAddressSequenceNumber, out var sequenceNumber))
-                    {
-                        if (untilSequenceNumber != -1 && sequenceNumber > untilSequenceNumber)
-                        {
-                            // Sequence numbers are monotonically increasing — stop processing this batch
-                            break;
-                        }
-                        aofProcessor.ProcessAofRecordInternal(virtualSublogIdx, entryPtr, payloadLength, true, out _, logAddressSequenceNumber);
-                        maxSequenceNumber = Math.Max(sequenceNumber, maxSequenceNumber);
-                    }
-                    entryLength += TsavoriteLog.UnsafeAlign(payloadLength);
-                }
-                else if (payloadLength < 0)
-                {
-                    // Only a single thread should commit metadata
-                    if (replayTaskIdx == 0)
-                    {
-                        TsavoriteLogRecoveryInfo info = new();
-                        info.Initialize(new ReadOnlySpan<byte>(ptr + entryLength, -payloadLength));
-                        replaySublog.UnsafeCommitMetadataOnly(info, isProtected);
-                    }
-                    entryLength += TsavoriteLog.UnsafeAlign(-payloadLength);
-                }
-                ptr += entryLength;
-            }
-
-            // Update max sequence number for this virtual sublog which is mapped
-            appendOnlyFile.readConsistencyManager.UpdateVirtualSublogMaxSequenceNumber(virtualSublogIdx, maxSequenceNumber);
+            replayTasks = [.. Enumerable.Range(0, serverOptions.AofReplayTaskCount).Select(i => Task.Run(() => RecoverReplayTaskAsync(i, physicalSublog)))];
         }
 
         public void Throttle() { }
@@ -285,7 +193,7 @@ namespace Garnet.server
         /// Starts a background task to replay and recover data until a specified address or when cancellation is requested.
         /// </summary>
         /// <returns>A Task representing the asynchronous recovery operation.</returns>
-        public Task CreateRecoverTaskAsync()
+        public Task RunAsync()
         {
             return Task.Run(async () =>
             {
